Rename llm package to prompt
