--- haystack/nodes/llm/__init__.py
@@ -1,1 +1,0 @@
-from haystack.nodes.llm.prompt_node import PromptNode, PromptTemplate, PromptModel
