+++ haystack/nodes/prompt/__init__.py
@@ -1,0 +1,1 @@
+from haystack.nodes.prompt.prompt_node import PromptNode, PromptTemplate, PromptModel
