+++ haystack/nodes/prompt/prompt_node.py
